login form and firebase authentication
