+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   google_fonts: ^3.0.1
   cloud_firestore: ^3.3.0
+  firebase_auth: ^3.5.0
 
 dev_dependencies:
   flutter_test:
